storage/requirements.txt
connexion[flask]
connexion[uvicorn]
swagger-ui-bundle
pyyaml
SQLAlchemy
mysqlclient
kafka-python